# server/swagger_server/models/job_info.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class JobInfo(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, job_id: int=None):  # noqa: E501
        """JobInfo - a model defined in Swagger

        :param job_id: The job_id of this JobInfo.  # noqa: E501
        :type job_id: int
        """
        self.swagger_types = {
            'job_id': int
        }

        self.attribute_map = {
            'job_id': 'job_id'
        }

        self._job_id = job_id

    @classmethod
    def from_dict(cls, dikt) -> 'JobInfo':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The JobInfo of this JobInfo.  # noqa: E501
        :rtype: JobInfo
        """
        return util.deserialize_model(dikt, cls)

    @property
    def job_id(self) -> int:
        """Gets the job_id of this JobInfo.


        :return: The job_id of this JobInfo.
        :rtype: int
        """
        return self._job_id

    @job_id.setter
    def job_id(self, job_id: int):
        """Sets the job_id of this JobInfo.


        :param job_id: The job_id of this JobInfo.
        :type job_id: int
        """

        self._job_id = job_id